+++ src/infra/redis/index.ts
@@ -1,2 +1,11 @@
+export { TypeGuard } from './guards/type.guard.js';
+export {
+	RedisAdapter,
+	StreamMessageReply,
+	StreamMessagesReply,
+	StreamNameClockPair,
+	Task,
+	XAutoClaimResponse,
+} from './interfaces/index.js';
 export * from './redis.factory.js';
 export * from './redis.module.js';
