+++ src/infra/y-redis/helper.ts
@@ -1,7 +1,6 @@
 import { RedisKey } from 'ioredis';
 import { array, map } from 'lib0';
-import { TypeGuard } from '../../infra/redis/guards/type.guard.js';
-import { StreamMessageReply, StreamMessagesReply } from '../../infra/redis/interfaces/index.js';
+import { StreamMessageReply, StreamMessagesReply, TypeGuard } from '../../infra/redis/index.js';
 import { YRedisMessage } from './interfaces/stream-message.js';
 
 /* This file contains the implementation of the functions,
