+++ src/infra/y-redis/testing/index.ts
@@ -1,0 +1,3 @@
+export { yRedisDocFactory } from './y-redis-doc.factory.js';
+export { yRedisMessageFactory } from './y-redis-message.factory.js';
+export { yRedisUserFactory } from './y-redis-user.factory.js';
