+++ src/modules/server/server.module.ts
@@ -1,14 +1,13 @@
 import { Module, ValidationPipe } from '@nestjs/common';
 import { APP_PIPE } from '@nestjs/core';
 import { App } from 'uWebSockets.js';
-import { AuthGuardModule } from '../../infra/auth-guard/auth-guard.module.js';
-import { AuthorizationModule } from '../../infra/authorization/authorization.module.js';
-import { ConfigurationModule } from '../../infra/configuration/configuration.module.js';
-import { LoggerModule } from '../../infra/logger/logger.module.js';
+import { AuthGuardModule } from '../../infra/auth-guard/index.js';
+import { AuthorizationModule } from '../../infra/authorization/index.js';
+import { ConfigurationModule } from '../../infra/configuration/index.js';
+import { LoggerModule } from '../../infra/logger/index.js';
 import { RedisModule } from '../../infra/redis/index.js';
-import { StorageModule } from '../../infra/storage/storage.module.js';
-import { YRedisClientModule } from '../../infra/y-redis/y-redis-client.module.js';
-import { YRedisServiceModule } from '../../infra/y-redis/y-redis-service.module.js';
+import { StorageModule } from '../../infra/storage/index.js';
+import { YRedisClientModule, YRedisServiceModule } from '../../infra/y-redis/index.js';
 import { TldrawConfigController } from './api/tldraw-confg.controller.js';
 import { TldrawDocumentController } from './api/tldraw-document.controller.js';
 import { WebsocketGateway } from './api/websocket.gateway.js';
