+++ src/modules/worker/worker.module.ts
@@ -1,9 +1,9 @@
 import { Module } from '@nestjs/common';
-import { ConfigurationModule } from '../../infra/configuration/configuration.module.js';
-import { LoggerModule } from '../../infra/logger/logger.module.js';
-import { RedisModule } from '../../infra/redis/redis.module.js';
-import { StorageModule } from '../../infra/storage/storage.module.js';
-import { YRedisClientModule } from '../../infra/y-redis/y-redis-client.module.js';
+import { ConfigurationModule } from '../../infra/configuration/index.js';
+import { LoggerModule } from '../../infra/logger/index.js';
+import { RedisModule } from '../../infra/redis/index.js';
+import { StorageModule } from '../../infra/storage/index.js';
+import { YRedisClientModule } from '../../infra/y-redis/index.js';
 import { WorkerConfig } from './worker.config.js';
 import { REDIS_FOR_WORKER } from './worker.const.js';
 import { WorkerService } from './worker.service.js';
